+++ functions/colors.py
@@ -12,7 +12,7 @@ from random import uniform
 
 from rpi_ws281x import *
 
-from control.light_wreath import wreath_setup
+from control.wreath import wreath_setup
 from functions.effects import clear
 from logger import LOGGER
